MVD2 Stop parsing on reaching range count. Some cleanup
